# Use sync.Map in the endpointregistry