--- a/routing/endpointregistry_test.go
+++ b/routing/endpointregistry_test.go
@@ -17,37 +17,50 @@ func TestEmptyRegistry(t *testing.T) {
 	m := r.GetMetrics("some key")
 
 	assert.Equal(t, time.Time{}, m.DetectedTime())
+	assert.Equal(t, time.Time{}, m.LastSeen())
 	assert.Equal(t, int64(0), m.InflightRequests())
 }
 
 func TestSetAndGet(t *testing.T) {
+	now := time.Now()
 	r := routing.NewEndpointRegistry(routing.RegistryOptions{})
 
 	mBefore := r.GetMetrics("some key")
-	r.IncInflightRequest("some key")
+	assert.Equal(t, time.Time{}, mBefore.DetectedTime())
+	assert.Equal(t, time.Time{}, mBefore.LastSeen())
+	assert.Equal(t, int64(0), mBefore.InflightRequests())
+
+	r.GetMetrics("some key").SetDetected(now.Add(-time.Second))
+	r.GetMetrics("some key").SetLastSeen(now)
+	r.GetMetrics("some key").IncInflightRequest()
 	mAfter := r.GetMetrics("some key")
 
-	assert.Equal(t, int64(0), mBefore.InflightRequests())
+	assert.Equal(t, now.Add(-time.Second), mBefore.DetectedTime())
+	assert.Equal(t, now, mBefore.LastSeen())
+	assert.Equal(t, int64(1), mBefore.InflightRequests())
+
+	assert.Equal(t, now.Add(-time.Second), mAfter.DetectedTime())
+	assert.Equal(t, now, mAfter.LastSeen())
 	assert.Equal(t, int64(1), mAfter.InflightRequests())
-
-	ts, _ := time.Parse(time.DateOnly, "2023-08-29")
-	mBefore = r.GetMetrics("some key")
-	r.SetDetectedTime("some key", ts)
-	mAfter = r.GetMetrics("some key")
-
-	assert.Equal(t, time.Time{}, mBefore.DetectedTime())
-	assert.Equal(t, ts, mAfter.DetectedTime())
 }
 
 func TestSetAndGetAnotherKey(t *testing.T) {
+	now := time.Now()
 	r := routing.NewEndpointRegistry(routing.RegistryOptions{})
 
-	r.IncInflightRequest("some key")
 	mToChange := r.GetMetrics("some key")
+	mToChange.IncInflightRequest()
+	mToChange.SetDetected(now.Add(-time.Second))
+	mToChange.SetLastSeen(now)
 	mConst := r.GetMetrics("another key")
 
 	assert.Equal(t, int64(0), mConst.InflightRequests())
+	assert.Equal(t, time.Time{}, mConst.DetectedTime())
+	assert.Equal(t, time.Time{}, mConst.LastSeen())
+
 	assert.Equal(t, int64(1), mToChange.InflightRequests())
+	assert.Equal(t, now.Add(-time.Second), mToChange.DetectedTime())
+	assert.Equal(t, now, mToChange.LastSeen())
 }
 
 func TestDoRemovesOldEntries(t *testing.T) {
@@ -73,9 +86,9 @@ func TestDoRemovesOldEntries(t *testing.T) {
 	assert.Equal(t, beginTestTs, mExist.DetectedTime())
 	assert.Equal(t, beginTestTs, mExistYet.DetectedTime())
 
-	r.IncInflightRequest("endpoint1.test:80")
-	r.IncInflightRequest("endpoint2.test:80")
-	r.DecInflightRequest("endpoint2.test:80")
+	mExist.IncInflightRequest()
+	mExistYet.IncInflightRequest()
+	mExistYet.DecInflightRequest()
 
 	routing.SetNow(r, func() time.Time {
 		return beginTestTs.Add(routing.ExportDefaultLastSeenTimeout + time.Second)
@@ -99,6 +112,95 @@ func TestDoRemovesOldEntries(t *testing.T) {
 
 	assert.Equal(t, time.Time{}, mRemoved.DetectedTime())
 	assert.Equal(t, int64(0), mRemoved.InflightRequests())
+}
+
+func TestMetricsMethodsDoNotAllocate(t *testing.T) {
+	r := routing.NewEndpointRegistry(routing.RegistryOptions{})
+	metrics := r.GetMetrics("some key")
+	now := time.Now()
+	metrics.SetDetected(now.Add(-time.Hour))
+	metrics.SetLastSeen(now)
+
+	allocs := testing.AllocsPerRun(100, func() {
+		assert.Equal(t, int64(0), metrics.InflightRequests())
+		metrics.IncInflightRequest()
+		assert.Equal(t, int64(1), metrics.InflightRequests())
+		metrics.DecInflightRequest()
+		assert.Equal(t, int64(0), metrics.InflightRequests())
+
+		metrics.DetectedTime()
+		metrics.LastSeen()
+	})
+	assert.Equal(t, now.Add(-time.Hour), metrics.DetectedTime())
+	assert.Equal(t, now, metrics.LastSeen())
+
+	assert.Equal(t, 0.0, allocs)
+}
+
+func TestRaceReadWrite(t *testing.T) {
+	r := routing.NewEndpointRegistry(routing.RegistryOptions{})
+	duration := time.Second
+
+	wg := sync.WaitGroup{}
+	wg.Add(2)
+	go func() {
+		defer wg.Done()
+		stop := time.After(duration)
+		for {
+			r.GetMetrics("some key")
+			select {
+			case <-stop:
+				return
+			default:
+			}
+		}
+	}()
+	go func() {
+		defer wg.Done()
+		stop := time.After(duration)
+		for {
+			r.GetMetrics("some key").IncInflightRequest()
+			select {
+			case <-stop:
+				return
+			default:
+			}
+		}
+	}()
+	wg.Wait()
+}
+
+func TestRaceTwoWriters(t *testing.T) {
+	r := routing.NewEndpointRegistry(routing.RegistryOptions{})
+	duration := time.Second
+
+	wg := sync.WaitGroup{}
+	wg.Add(2)
+	go func() {
+		defer wg.Done()
+		stop := time.After(duration)
+		for {
+			r.GetMetrics("some key").IncInflightRequest()
+			select {
+			case <-stop:
+				return
+			default:
+			}
+		}
+	}()
+	go func() {
+		defer wg.Done()
+		stop := time.After(duration)
+		for {
+			r.GetMetrics("some key").DecInflightRequest()
+			select {
+			case <-stop:
+				return
+			default:
+			}
+		}
+	}()
+	wg.Wait()
 }
 
 func printTotalMutexWaitTime(b *testing.B) {
@@ -133,20 +235,24 @@ func benchmarkIncInflightRequests(b *testing.B, name string, goroutines int) {
 
 	b.Run(name, func(b *testing.B) {
 		r := routing.NewEndpointRegistry(routing.RegistryOptions{})
+		now := time.Now()
+
 		for i := 1; i < mapSize; i++ {
-			r.IncInflightRequest(fmt.Sprintf("foo-%d", i))
+			r.GetMetrics(fmt.Sprintf("foo-%d", i)).IncInflightRequest()
 		}
-		r.IncInflightRequest(key)
-		r.IncInflightRequest(key)
+		r.GetMetrics(key).IncInflightRequest()
+		r.GetMetrics(key).SetDetected(now)
 
 		wg := sync.WaitGroup{}
 		b.ResetTimer()
+		b.ReportAllocs()
 		for i := 0; i < goroutines; i++ {
 			wg.Add(1)
 			go func() {
 				defer wg.Done()
+				metrics := r.GetMetrics(key)
 				for n := 0; n < b.N/goroutines; n++ {
-					r.IncInflightRequest(key)
+					metrics.IncInflightRequest()
 				}
 			}()
 		}
@@ -169,21 +275,24 @@ func benchmarkGetInflightRequests(b *testing.B, name string, goroutines int) {
 
 	b.Run(name, func(b *testing.B) {
 		r := routing.NewEndpointRegistry(routing.RegistryOptions{})
+		now := time.Now()
 		for i := 1; i < mapSize; i++ {
-			r.IncInflightRequest(fmt.Sprintf("foo-%d", i))
+			r.GetMetrics(fmt.Sprintf("foo-%d", i)).IncInflightRequest()
 		}
-		r.IncInflightRequest(key)
-		r.IncInflightRequest(key)
+		r.GetMetrics(key).IncInflightRequest()
+		r.GetMetrics(key).SetDetected(now)
 
 		var dummy int64
 		wg := sync.WaitGroup{}
 		b.ResetTimer()
+		b.ReportAllocs()
 		for i := 0; i < goroutines; i++ {
 			wg.Add(1)
 			go func() {
 				defer wg.Done()
+				metrics := r.GetMetrics(key)
 				for n := 0; n < b.N/goroutines; n++ {
-					dummy = r.GetMetrics(key).InflightRequests()
+					dummy = metrics.InflightRequests()
 				}
 			}()
 		}
@@ -198,5 +307,46 @@ func BenchmarkGetInflightRequests(b *testing.B) {
 	goroutinesNums := []int{1, 2, 3, 4, 5, 6, 7, 8, 12, 16, 24, 32, 48, 64, 128, 256, 512, 768, 1024, 1536, 2048, 4096}
 	for _, goroutines := range goroutinesNums {
 		benchmarkGetInflightRequests(b, fmt.Sprintf("%d goroutines", goroutines), goroutines)
+	}
+}
+
+func benchmarkGetDetectedTime(b *testing.B, name string, goroutines int) {
+	const key string = "some key"
+	const mapSize int = 10000
+
+	b.Run(name, func(b *testing.B) {
+		r := routing.NewEndpointRegistry(routing.RegistryOptions{})
+		now := time.Now()
+		for i := 1; i < mapSize; i++ {
+			r.GetMetrics(fmt.Sprintf("foo-%d", i)).IncInflightRequest()
+		}
+		r.GetMetrics(key).IncInflightRequest()
+		r.GetMetrics(key).SetDetected(now)
+
+		var dummy time.Time
+		wg := sync.WaitGroup{}
+		b.ResetTimer()
+		b.ReportAllocs()
+		for i := 0; i < goroutines; i++ {
+			wg.Add(1)
+			go func() {
+				defer wg.Done()
+				metrics := r.GetMetrics(key)
+				for n := 0; n < b.N/goroutines; n++ {
+					dummy = metrics.DetectedTime()
+				}
+			}()
+		}
+		dummy = dummy.Add(time.Second)
+		wg.Wait()
+
+		printTotalMutexWaitTime(b)
+	})
+}
+
+func BenchmarkGetDetectedTime(b *testing.B) {
+	goroutinesNums := []int{1, 2, 3, 4, 5, 6, 7, 8, 12, 16, 24, 32, 48, 64, 128, 256, 512, 768, 1024, 1536, 2048, 4096}
+	for _, goroutines := range goroutinesNums {
+		benchmarkGetDetectedTime(b, fmt.Sprintf("%d goroutines", goroutines), goroutines)
 	}
 }