@AlexanderYastrebov 2023-12-14
Please add a benchmark showing performance and allocations of reading metrics.