@RomanZavodskikh 2023-12-14
@AlexanderYastrebov I posted the benchmark inside PR description.
tl;dr It's good as far as user does not call `GetMetrics` too much.